Бургер кафе в отдельный файл. Добавлен роутер и первый маршрут.
